changed attack particel from sprite to particle system, made a parent class for all enemies
