+++ Assets/Resources/Prefabs/Particle Effects.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 836db90baed55bc41a62f6fe6160fbb5
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
