+++ Assets/Resources/Prefabs/Particle Effects/HitEffect1.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: bb916d6a8cad70d4fab2ce9017edede7
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
