+++ Assets/Scripts/Animators/SelfDestructParticle.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 7b5fe293265b787419d067de6b6f556d
