+++ Assets/Scripts/Enemy/RangedEnemyController.cs
@@ -3,34 +3,10 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-public class RangedEnemyController : MonoBehaviour
+public class RangedEnemyController : EnemyControllerBase
 {
-    public int health = 10;
-    
-    // Start is called before the first frame update
-    void Start()
+    protected override Vector3 GetDirection()
     {
-        
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-        if (health <= 0)
-        {
-            Destroy(gameObject);
-        }
-    }
-
-    // reduce health on damage from a PlayerAttack
-    private void OnTriggerEnter2D(Collider2D collision)
-    {
-        if (collision.gameObject.CompareTag("DashAttack"))
-        {
-            health -= (int)PlayerStats.Instance.GetDashDamage();
-        } else if (collision.gameObject.CompareTag("PlayerAttack"))
-        {
-            health -= (int)PlayerStats.Instance.GetAttackDamage();
-        }
+        return Vector3.zero;
     }
 }
